RUMM-1161 use TestConfiguration for Looper.getMainLooper
