+++ dd-sdk-android/src/test/kotlin/com/datadog/android/tracing/internal/AndroidTracerTest.kt
@@ -15,12 +15,11 @@ import com.datadog.android.log.LogAttributes
 import com.datadog.android.rum.internal.RumFeature
 import com.datadog.android.tracing.AndroidTracer
 import com.datadog.android.utils.config.GlobalRumMonitorTestConfiguration
-import com.datadog.android.utils.disposeMainLooper
+import com.datadog.android.utils.config.MainLooperTestConfiguration
 import com.datadog.android.utils.forge.Configurator
 import com.datadog.android.utils.mockContext
 import com.datadog.android.utils.mockCoreFeature
 import com.datadog.android.utils.mockDevLogHandler
-import com.datadog.android.utils.prepareMainLooper
 import com.datadog.opentracing.DDSpan
 import com.datadog.opentracing.LogHandler
 import com.datadog.opentracing.scopemanager.ContextualScopeManager
@@ -65,7 +64,6 @@ import org.mockito.quality.Strictness
     ExtendWith(ForgeExtension::class),
     ExtendWith(TestConfigurationExtension::class)
 )
-
 @MockitoSettings(strictness = Strictness.LENIENT)
 @ForgeConfiguration(Configurator::class)
 internal class AndroidTracerTest {
@@ -91,7 +89,6 @@ internal class AndroidTracerTest {
         fakeToken = forge.anHexadecimalString()
         mockAppContext = mockContext()
         mockCoreFeature()
-        prepareMainLooper()
         TracesFeature.initialize(mockAppContext, Configuration.DEFAULT_TRACING_CONFIG)
         RumFeature.initialize(mockAppContext, Configuration.DEFAULT_RUM_CONFIG)
         testedTracerBuilder = AndroidTracer.Builder()
@@ -114,7 +111,6 @@ internal class AndroidTracerTest {
             ContextualScopeManager::class.java.getStaticValue("tlsScope")
         tlsScope.remove()
         RumFeature.stop()
-        disposeMainLooper()
     }
 
     // region Tracer
@@ -409,11 +405,12 @@ internal class AndroidTracerTest {
 
     companion object {
         val rumMonitor = GlobalRumMonitorTestConfiguration()
+        val mainLooper = MainLooperTestConfiguration()
 
         @TestConfigurationsProvider
         @JvmStatic
         fun getTestConfigurations(): List<TestConfiguration> {
-            return listOf(rumMonitor)
+            return listOf(rumMonitor, mainLooper)
         }
     }
 }
