+++ dd-sdk-android/src/test/kotlin/com/datadog/android/utils/DatadogExt.kt
@@ -11,7 +11,6 @@ import android.content.pm.ApplicationInfo
 import android.content.pm.PackageInfo
 import android.content.pm.PackageManager
 import android.os.Build
-import android.os.Looper
 import com.datadog.android.BuildConfig
 import com.datadog.android.Datadog
 import com.datadog.android.core.internal.CoreFeature
@@ -21,9 +20,7 @@ import com.datadog.android.log.Logger
 import com.datadog.android.log.internal.logger.LogHandler
 import com.datadog.android.privacy.TrackingConsent
 import com.datadog.tools.unit.getFieldValue
-import com.datadog.tools.unit.getStaticValue
 import com.datadog.tools.unit.setFieldValue
-import com.datadog.tools.unit.setStaticValue
 import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.doReturn
 import com.nhaarman.mockitokotlin2.mock
@@ -130,24 +127,4 @@ internal fun mockCoreFeature(
     CoreFeature.trackingConsentProvider = mock()
     CoreFeature.userInfoProvider = mock()
     whenever(CoreFeature.trackingConsentProvider.getConsent()) doReturn trackingConsent
-}
-
-internal fun prepareMainLooper() {
-    val mainLooper = Looper.getMainLooper()
-    if (mainLooper == null) {
-        try {
-            @Suppress("DEPRECATION")
-            Looper.prepareMainLooper()
-        } catch (e: IllegalStateException) {
-            // main looper already prepared
-        }
-    }
-    checkNotNull(Looper.getMainLooper())
-}
-
-internal fun disposeMainLooper() {
-    Looper::class.java.setStaticValue("sMainLooper", null)
-    Looper::class.java.getStaticValue<Looper, ThreadLocal<Looper>>("sThreadLocal").set(null)
-
-    check(Looper.getMainLooper() == null)
 }
